Mark incident metadata endpoint as deprecated
.. and remove it from the nascent API V2.
